requirements.txt
pandas
numpy
matplotlib
nltk
spacy
scikit-learn
sumy
rouge-score
transformers
torch
tqdm
rich
pyyaml 
